+++ metadata.rb
@@ -4,5 +4,5 @@ maintainer_email 'radek@opsline.com'
 license          'All rights reserved'
 description      'Helper recipes for chef client'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.3.0'
+version          '0.4.0'